Initial tool for creating Amlogic Meson GXBB boot images
This simplistic tool does almost the same as:

  aml_encrypt_gxb --bootsig --input boot_new.bin --output test.out

Differences are the 16 initial bytes, as well as 4 bytes in the header
of each @AML header that unfortunately affect their SHA-256 checksums -
it seems to work nontheless, just looks ugly in a hex diff.

Little Endian only for now.

Unimplemented are --amluserkey and --efuse, as they seem not needed on
Odroid-C2.

Odroid-C2 discards the first 96 512-byte blocks of our output (0xC000),
so a separate function do_fip() is factored out for later possibly
skipping BL2 processing.

Read the code before use and USE AT YOUR OWN RISK!

Signed-off-by: Andreas Färber <andreas.faerber@web.de>