Add headers with license information
Source code is licensed as GPLv2 "or (at your option) any later version".
The aim is to have an Open Source license that 1) assures any additions
or fixes can find their way back into the main tool, 2) is compatible
with GPLv2-only projects such as U-Boot or Linux, 3) allows easy
redistribution under the simplified terms of the GPL-3.0+.

Headers are dual-licensed as GPLv2+ and as MIT, to allow maximum reuse
and collaboration on the core data structures. Future headers that apply
only to the GPL'ed tools do not need to be dual-licensed.

This resolves issue #1.

Signed-off-by: Andreas Färber <andreas.faerber@web.de>